Refactor signup and login further
